集成nacos，配合MinIo完成媒资管理模块开发，下一步进行视频上传，将视频文件上传至分布式文件存储系统，并实现断点续传的功能，保证上传文件体验感。

+++ xuecheng-plus-content/xuecheng-plus-content-model/src/main/java/com/tunan/content/model/dto/CourseBaseInfoDto.java
@@ -3,10 +3,12 @@ package com.tunan.content.model.dto;
 import com.tunan.content.model.po.CourseBase;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @description 课程基本信息dto
- * @author Mr.M
- * @date 2022/9/7 17:44
+ * @author Jack
+ * @date 2023/9/19 17:44
  * @version 1.0
  */
 @Data
@@ -21,7 +23,7 @@ public class CourseBaseInfoDto extends CourseBase {
  /**
   * 价格
   */
- private Float price;
+ private BigDecimal price;
 
 
  /**
